TestingPower -> BrowserEfficiencyTest
Changed "TestingPower" to "BrowserEfficiencyTest" to reflect the ability
of BET to do more than power now
