+++ BrowserEfficiencyTest.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+VisualStudioVersion = 14.0.25123.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingPower", "TestingPower\TestingPower.csproj", "{F54DFCBE-CD4C-465A-8523-F1C79C3E05C4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BrowserEfficiencyTest", "BrowserEfficiencyTest\BrowserEfficiencyTest.csproj", "{F54DFCBE-CD4C-465A-8523-F1C79C3E05C4}"
 	ProjectSection(ProjectDependencies) = postProject
 		{8F9D79CD-9E21-464F-80D6-4B502AA51845} = {8F9D79CD-9E21-464F-80D6-4B502AA51845}
 	EndProjectSection
