+++ BrowserEfficiencyTest/BrowserEfficiencyTest.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{F54DFCBE-CD4C-465A-8523-F1C79C3E05C4}</ProjectGuid>
     <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>TestingPower</RootNamespace>
-    <AssemblyName>TestingPower</AssemblyName>
+    <RootNamespace>BrowserEfficiencyTest</RootNamespace>
+    <AssemblyName>BrowserEfficiencyTest</AssemblyName>
     <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
