+++ BrowserEfficiencyTest/Scenarios/RedditSearchSubreddit.cs
@@ -1,6 +1,6 @@
-﻿//--------------------------------------------------------------
+//--------------------------------------------------------------
 //
-// Microsoft Edge Power Test
+// Browser Efficiency Test
 // Copyright(c) Microsoft Corporation
 // All rights reserved.
 //
@@ -25,29 +25,50 @@
 //
 //--------------------------------------------------------------
 
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using OpenQA.Selenium.Remote;
+using System.Threading;
+using OpenQA.Selenium;
 
-namespace TestingPower
+namespace BrowserEfficiencyTest
 {
-    /// <summary>
-    /// This scenario is designed to end quickly, and is for testing
-    /// </summary>
-    internal class FastScenario : Scenario
+    internal class RedditSearchSubreddit : Scenario
     {
-        public FastScenario()
+        public RedditSearchSubreddit()
         {
-            Name = "fastScenario";
-            Duration = 10;
+            Name = "reddit";
         }
-
         public override void Run(RemoteWebDriver driver, string browser, List<UserInfo> logins)
         {
-            driver.Navigate().GoToUrl("http://www.google.com");
+            driver.Navigate().GoToUrl("https://www.reddit.com/");
+            Thread.Sleep(2000);
+
+            const string SearchTerm = "marvel contest";
+            var search = driver.FindElementByName("q");
+
+            search.SendKeys(SearchTerm);
+            search.SendKeys(Keys.Tab);
+            search.SendKeys(Keys.Enter);
+
+            while (!driver.Title.Contains(SearchTerm))
+            {
+                Thread.Sleep(1000);
+            }
+
+            Thread.Sleep(1000);
+            var subreddit = driver.FindElementByClassName("search-title");
+            driver.Navigate().GoToUrl("https://www.reddit.com/r/ContestOfChampions/?ref=search_subreddits");
+            Thread.Sleep(5000);
+
+            driver.ScrollPage(3);
+
+            var subredditThread = driver.FindElementByClassName("thumbnail");
+
+            Thread.Sleep(5000);
+            driver.Navigate().GoToUrl("https://www.reddit.com/r/ContestOfChampions/comments/4luknw/rank_upteam_buildingawakening_post/");
+            Thread.Sleep(10000);
+
+            driver.ScrollPage(15);
         }
     }
 }
