+++ BrowserEfficiencyTest/Scenarios/YoutubeWatchVideo.cs
@@ -1,6 +1,6 @@
 //--------------------------------------------------------------
 //
-// Microsoft Edge Power Test
+// Browser Efficiency Test
 // Copyright(c) Microsoft Corporation
 // All rights reserved.
 //
@@ -29,19 +29,35 @@ using OpenQA.Selenium.Remote;
 using System.Collections.Generic;
 using System.Threading;
 
-namespace TestingPower
+namespace BrowserEfficiencyTest
 {
-    internal class Msn : Scenario
+    internal class YoutubeWatchVideo : Scenario
     {
-        public Msn()
+        public YoutubeWatchVideo()
         {
-            this.Name = "msn";
+            this.Name = "youtube";
+            // Leave the default time
         }
 
         public override void Run(RemoteWebDriver driver, string browser, List<UserInfo> logins)
         {
-            driver.Navigate().GoToUrl("http://www.msn.com");
-            Thread.Sleep(1000);
+            // Browse to Youtube if we're not there already
+            if (!driver.Url.Contains("youtube.com"))
+            {
+                driver.Navigate().GoToUrl("https://www.youtube.com/watch?v=l42U5Cwn1Y0");
+                Thread.Sleep(2000);
+            }
+
+            string movieId = "movie_player";
+            var player = driver.FindElementById(movieId);
+
+            // Play if it's paused
+            if (player.GetAttribute("class").Contains("paused-mode"))
+            {
+                player.Click();
+            }
+
+            Thread.Sleep(2000);
         }
     }
 }
