+++ BrowserEfficiencyTest/UserInfo.cs
@@ -1,6 +1,6 @@
-//--------------------------------------------------------------
+﻿//--------------------------------------------------------------
 //
-// Microsoft Edge Power Test
+// MBrowser Efficiency Test
 // Copyright(c) Microsoft Corporation
 // All rights reserved.
 //
@@ -25,37 +25,19 @@
 //
 //--------------------------------------------------------------
 
-using System.Collections.Generic;
-using System.Threading;
-using OpenQA.Selenium.Remote;
-using OpenQA.Selenium;
+using Newtonsoft.Json;
 
-namespace TestingPower
+namespace BrowserEfficiencyTest
 {
-    internal class GoogleSearch : Scenario
+    internal class UserInfo
     {
-        public GoogleSearch()
-        {
-            Name = "google";
-            // Default time
-        }
+        [JsonProperty("Domain")]
+        public string Domain { get; set; }
 
-        public override void Run(RemoteWebDriver driver, string browser, List<UserInfo> logins)
-        {
-            driver.Navigate().GoToUrl("http://www.google.com");
+        [JsonProperty("UserName")]
+        public string UserName { get; set; }
 
-            Thread.Sleep(5 * 1000);
-
-            // Search for "Seattle" and hit enter
-            var searchBox = driver.FindElementByXPath("//*[@title='Search']");
-            foreach (char c in "Seattle")
-            {
-                searchBox.SendKeys(c.ToString());
-                Thread.Sleep(75);
-            }
-            driver.Keyboard.SendKeys(Keys.Enter);
-
-            // Simply yield control back to the main thread and look at results
-        }
+        [JsonProperty("Password")]
+        public string PassWord { get; set; }
     }
 }
